backend/pyproject.toml
[tool.poetry]
name = "socilink"
version = "0.1.0"
description = ""
authors = ["daniel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
channels = "^4.0.0"
pillow = "^10.0.0"
daphne = "^4.0.0"
channels-redis = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
